+++ src/main/java/com/epam/ta/reportportal/core/imprt/impl/junit/XunitImportHandler.java
@@ -34,7 +34,9 @@ import org.xml.sax.Attributes;
 import org.xml.sax.SAXException;
 import org.xml.sax.helpers.DefaultHandler;
 
+import java.time.Instant;
 import java.time.LocalDateTime;
+import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeFormatterBuilder;
 import java.time.temporal.ChronoUnit;
@@ -91,12 +93,15 @@ public class XunitImportHandler extends DefaultHandler {
         switch (XunitReportTag.fromString(qName)) {
             case TESTSUITE:
                 startRootItem(attributes.getValue(XunitReportTag.ATTR_NAME.getValue()),
-                        attributes.getValue(XunitReportTag.TIMESTAMP.getValue()));
+                        attributes.getValue(XunitReportTag.TIMESTAMP.getValue())
+                );
                 break;
             case TESTCASE:
                 startTestItem(attributes.getValue(XunitReportTag.ATTR_NAME.getValue()),
-                        attributes.getValue(XunitReportTag.ATTR_TIME.getValue()));
+                        attributes.getValue(XunitReportTag.ATTR_TIME.getValue())
+                );
                 break;
+            case ERROR:
             case FAILURE:
                 message = new StringBuilder();
                 status = Status.FAILED;
@@ -123,6 +128,7 @@ public class XunitImportHandler extends DefaultHandler {
             case SKIPPED:
                 attachLog(LogLevel.ERROR);
                 break;
+            case ERROR:
             case FAILURE:
                 attachLog(LogLevel.ERROR);
                 break;
@@ -143,11 +149,7 @@ public class XunitImportHandler extends DefaultHandler {
 
     private void startRootItem(String name, String timestamp) {
         if (null != timestamp) {
-            DateTimeFormatter formatter = new DateTimeFormatterBuilder()
-                    .append(DateTimeFormatter.ISO_LOCAL_DATE_TIME)
-                    .optionalStart().appendZoneId().optionalEnd()
-                    .toFormatter();
-            startItemTime = LocalDateTime.parse(timestamp, formatter);
+            startItemTime = parseTimeStamp(timestamp);
             if (startSuiteTime.isAfter(startItemTime)) {
                 startSuiteTime = LocalDateTime.of(startItemTime.toLocalDate(), startItemTime.toLocalTime());
             }
@@ -163,6 +165,25 @@ public class XunitImportHandler extends DefaultHandler {
         itemsIds.push(id);
     }
 
+    private LocalDateTime parseTimeStamp(String timestamp) {
+        LocalDateTime localDateTime = null;
+        try {
+            localDateTime = LocalDateTime.ofInstant(Instant.ofEpochMilli(Long.parseLong(timestamp)), ZoneId.systemDefault());
+        } catch (NumberFormatException e) {
+            //ignored
+        }
+        if (null == localDateTime) {
+            DateTimeFormatter formatter = new DateTimeFormatterBuilder().appendOptional(DateTimeFormatter.RFC_1123_DATE_TIME)
+                    .appendOptional(DateTimeFormatter.ISO_LOCAL_DATE_TIME)
+                    .optionalStart()
+                    .appendZoneId()
+                    .optionalEnd()
+                    .toFormatter();
+            localDateTime = LocalDateTime.parse(timestamp, formatter);
+        }
+        return localDateTime;
+    }
+
     private void startTestItem(String name, String duration) {
         StartTestItemRQ rq = new StartTestItemRQ();
         rq.setLaunchId(launchId);
@@ -171,6 +192,7 @@ public class XunitImportHandler extends DefaultHandler {
         rq.setName(name);
         String id = startTestItemHandler.startChildItem(projectId, rq, itemsIds.peekLast()).getId();
         currentDuration = toMillis(duration);
+        currentId = id;
         itemsIds.push(id);
     }
 
@@ -198,7 +220,7 @@ public class XunitImportHandler extends DefaultHandler {
             SaveLogRQ saveLogRQ = new SaveLogRQ();
             saveLogRQ.setLevel(logLevel.name());
             saveLogRQ.setLogTime(toDate(startItemTime));
-            saveLogRQ.setMessage(message.toString());
+            saveLogRQ.setMessage(message.toString().trim());
             saveLogRQ.setTestItemId(currentId);
             createLogHandler.createLog(saveLogRQ, null, projectId);
         }
@@ -209,7 +231,7 @@ public class XunitImportHandler extends DefaultHandler {
             SaveLogRQ saveLogRQ = new SaveLogRQ();
             saveLogRQ.setLevel(logLevel.name());
             saveLogRQ.setLogTime(toDate(startItemTime));
-            saveLogRQ.setMessage(message.toString());
+            saveLogRQ.setMessage(message.toString().trim());
             saveLogRQ.setTestItemId(itemsIds.getFirst());
             createLogHandler.createLog(saveLogRQ, null, projectId);
         }